refactor: remove exports from package.json and adjust imports
Exports makes it hard to access other paths in the project
Until we arrive on stable interfaces to access other parts of git-proxy
removing exports would be helpful.
Import paths have been adjusted slightly for plugins and the docs for
plugins due to the exports aliases being a bit different.