fix: make loadFromGit non-interactive
loadFromGit should be sufficiently authenticated and if it isn't it's
not expected to get interactive input. Inform git not to wait for
interactive action.
This also resolves an issue in the tests where "should throw error
if repository is valid URL but not a git repository" test times out
waiting for a username.